fix bugs and add story mode menu

+++ Assets/Scripts/LevelEditor/Menu/SaveAndLoad/SaveLevel.cs
@@ -4,52 +4,69 @@ using System.IO;
 using UnityEngine;
 using UnityEngine.UI;
 
-using UnityEditor;
+// using UnityEditor;
 
 public class SaveLevel : MonoBehaviour {
 
-    public void SaveAs () {
+    public void Save () {
 
         byte[] levelData = FindObjectOfType<LevelDeconstructor>().Deconstruct();
 
         PerformSave(levelData);
     }
 
-    public string saveCache = "";
-
-    public void Save () {
-
-        if (saveCache == "") SaveAs();
-        else {
-
-            byte[] levelData = FindObjectOfType<LevelDeconstructor>().Deconstruct();
-            
-            File.WriteAllBytes(saveCache, levelData);
-        }
-    }
+    public InputField nameField;
 
     public void PerformSave (byte[] levelData) {
 
-        string savePath = EditorUtility.SaveFilePanel(
-            "Save Level",
-            Application.persistentDataPath + "/Levels",
-            "My Awesome Level.chm",
-            "chm"
-        );
+        string savePath = nameField.text;
+
+        savePath = savePath.Replace('/', '_');
+        savePath = savePath.Replace('\\', '_');
+        savePath = savePath.Replace('\'', '_');
+        savePath = savePath.Replace('\"', '_');
+        savePath = savePath.Replace('~', '_');
+        savePath = savePath.Replace('!', '_');
+        savePath = savePath.Replace('@', '_');
+        savePath = savePath.Replace('$', '_');
+        savePath = savePath.Replace('%', '_');
+        savePath = savePath.Replace('*', '_');
+        savePath = savePath.Replace(';', '_');
+        savePath = savePath.Replace(':', '_');
+
+        savePath = Application.persistentDataPath + "/Levels/" + savePath + ".chm";
+
+        // string savePath = EditorUtility.SaveFilePanel(
+        //     "Save Level",
+        //     Application.persistentDataPath + "/Levels",
+        //     "My Awesome Level.chm",
+        //     "chm"
+        // );
 
         if (savePath.Length == 0) {
 
-            EditorUtility.DisplayDialog(
-                "Something went wrong!",
-                "You must enter a location to save the Level!",
-                "Understood."
-            );
+            PopupManager.Popup("Failed To Save!", "Filename was null.");
+
+            // EditorUtility.DisplayDialog(
+            //     "Something went wrong!",
+            //     "You must enter a location to save the Level!",
+            //     "Understood."
+            // );
 
             return;
         }
 
+        if (File.Exists(savePath)) {
+
+            PopupManager.Popup("Failed To Save!", "Duplicate file found. Please try a different name.");
+            return;
+        }
+
+        if (!Directory.Exists(Application.persistentDataPath + "/Levels"))
+            Directory.CreateDirectory(Application.persistentDataPath + "/Levels");
+
         File.WriteAllBytes(savePath, levelData);
 
-        saveCache = savePath;
+        nameField.text = "";
     }
 }
